fix: unify selected chat provider runtime state
