+++ src/agent/useAgent.ts
@@ -2,8 +2,8 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import { resolveAgentReferences } from './context';
 import type { AgentContext, AgentReference } from './context';
 import type { AgentRuntimeModule, LLMMessage } from './runtime';
-import { normalizeLlmProvider, MODEL, PROVIDER } from './providerConfig';
-import type { LlmProvider } from './providerConfig';
+import { normalizeLlmProvider, MODEL, PROVIDER } from './client';
+import type { LlmProvider } from './client';
 import { normalizeLlmMessages, prepareMessagesForProvider } from './messages';
 import { makeDraft, replayActions } from '../editor/store';
 import { buildOperation, buildProposal, isProposalStale, partitionProposalActions, type Operation, type Proposal } from './proposal';
